# API-Gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    description: Расширенный ответ с ошибкой
    properties:
      code:
        example: INVALID_INPUT
        type: string
      details:
        items:
          $ref: '#/definitions/models.ValidationError'
        type: array
      error:
        example: Bad Request
        type: string
      message:
        example: Invalid input data
        type: string
    type: object
  models.FileInfo:
    properties:
      category:
        description: '"avatar", "resume", "logo", etc.'
        type: string
      direct_url:
        description: Прямая ссылка через бекенд
        type: string
      id:
        type: string
      name:
        type: string
      type:
        description: '"image", "document", "other"'
        type: string
      url:
        description: Presigned URL для фронта
        type: string
    type: object
  models.FileUploadResponse:
    properties:
      file_info:
        $ref: '#/definitions/models.FileInfo'
      message:
        type: string
    type: object
...
  models.PaginationResponse:
    properties:
      current_page:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
...
  models.Vacancy:
    properties:
      attachment_id:
        type: string
      attachment_url:
        type: string
      company_id:
        type: string
      create_at:
        type: string
      experience:
        type: integer
      id:
        type: string
      position_status:
        type: string
      salary:
        type: integer
      schedule:
        type: string
      title:
        type: string
      work_format:
        type: string
    type: object
  models.VacancyList:
    properties:
      pagination:
        $ref: '#/definitions/models.PaginationResponse'
      vacancies:
        items:
          $ref: '#/definitions/models.Vacancy'
        type: array
    type: object
  models.ValidationError:
    description: Детализированная ошибка валидации
    properties:
      code:
        example: REQUIRED_FIELD
        type: string
      field:
        example: email
        type: string
      message:
        example: Email is required
        type: string
    type: object
host: localhost:8000
info:
  contact:
    email: support@studjobs.ru
    name: API Support
  description: API Gateway для платформы StudJobs HH
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: StudJobs HH API Gateway
  version: "1.0"
paths:
...
  /hr/vacancy:
    get:
      consumes:
      - application/json
      description: Возвращает список вакансий для текущего HR с поддержкой фильтрации
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Количество элементов на странице
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Фильтр по ID компании
        in: query
        name: company_id
        type: string
      - description: Фильтр по статусу позиции
        in: query
        name: position_status
        type: string
      - description: Фильтр по формату работы
        in: query
        name: work_format
        type: string
      - description: Фильтр по графику работы
        in: query
        name: schedule
        type: string
      - description: Минимальная зарплата
        in: query
        name: min_salary
        type: integer
      - description: Максимальная зарплата
        in: query
        name: max_salary
        type: integer
      - description: Минимальный опыт работы (лет)
        in: query
        name: min_experience
        type: integer
      - description: Максимальный опыт работы (лет)
        in: query
        name: max_experience
        type: integer
      - description: Поиск по названию вакансии
        in: query
        name: search_title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список вакансий HR
          schema:
            $ref: '#/definitions/models.VacancyList'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить вакансии HR
      tags:
      - Vacancies
    post:
      consumes:
      - application/json
      description: Создает новую вакансию для текущего HR
      parameters:
      - description: Данные вакансии
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Vacancy'
      produces:
      - application/json
      responses:
        "201":
          description: Созданная вакансия
          schema:
            $ref: '#/definitions/models.Vacancy'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать вакансию
      tags:
      - Vacancies
  /hr/vacancy/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет вакансию текущего HR
      parameters:
      - description: ID вакансии
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Вакансия не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить вакансию
      tags:
      - Vacancies
    patch:
      consumes:
      - application/json
      description: Обновляет данные вакансии текущего HR
      parameters:
      - description: ID вакансии
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Vacancy'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная вакансия
          schema:
            $ref: '#/definitions/models.Vacancy'
        "400":
          description: Неверные данные запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Вакансия не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить вакансию
      tags:
      - Vacancies
  /positions:
    get:
      consumes:
      - application/json
      description: Возвращает список всех доступных должностей/позиций
      produces:
      - application/json
      responses:
        "200":
          description: Список позиций
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список позиций
      tags:
      - Vacancies
...
  /vacancy:
    get:
      consumes:
      - application/json
      description: Возвращает список вакансий с поддержкой пагинации и фильтрации
      parameters:
      - default: 1
        description: Номер страницы
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Количество элементов на странице
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Фильтр по ID компании
        in: query
        name: company_id
        type: string
      - description: Фильтр по статусу позиции
        in: query
        name: position_status
        type: string
      - description: Фильтр по формату работы
        in: query
        name: work_format
        type: string
      - description: Фильтр по графику работы
        in: query
        name: schedule
        type: string
      - description: Минимальная зарплата
        in: query
        name: min_salary
        type: integer
      - description: Максимальная зарплата
        in: query
        name: max_salary
        type: integer
      - description: Минимальный опыт работы (лет)
        in: query
        name: min_experience
        type: integer
      - description: Максимальный опыт работы (лет)
        in: query
        name: max_experience
        type: integer
      - description: Поиск по названию вакансии
        in: query
        name: search_title
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список вакансий
          schema:
            $ref: '#/definitions/models.VacancyList'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить список вакансий
      tags:
      - Vacancies
  /vacancy/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает информацию о вакансии по указанному идентификатору
      parameters:
      - description: ID вакансии
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Информация о вакансии
          schema:
            $ref: '#/definitions/models.Vacancy'
        "400":
          description: Неверный ID вакансии
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Вакансия не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить вакансию по ID
      tags:
      - Vacancies
  /vacancy/{id}/files/attachment:
    delete:
      consumes:
      - application/json
      description: Удаляет прикрепленный файл вакансии
      parameters:
      - description: ID вакансии
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Сообщение об успешном удалении
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: У вакансии нет вложения
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Вакансия не найдена
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить вложение вакансии
      tags:
      - Vacancies
    post:
      consumes:
      - multipart/form-data
      description: Загружает прикрепленный файл для вакансии
      parameters:
      - description: ID вакансии
        in: path
        name: id
        required: true
        type: string
      - description: Файл вложения (макс. 10MB)
        in: formData
        name: attachment
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Информация о загруженном файле
          schema:
            $ref: '#/definitions/models.FileUploadResponse'
        "400":
          description: Неверный запрос или файл слишком большой
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Неавторизованный доступ
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Загрузить вложение вакансии
      tags:
      - Vacancies
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
  RoleAuth:
    description: Роль пользователя
    in: header
    name: X-User-Role
    type: apiKey
swagger: "2.0"